budget_bot/requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
aiosqlite==0.19.0
